Move bored.py to the same folder as other functions
