+++ PyChan/Core/Commands/Text/Functions/bored.py
