+++ PyChan/Core/Commands/Text/text.py
@@ -1,6 +1,6 @@
 import nextcord
 from nextcord.ext import commands
-from Core.Commands.Text.Functions.bored.bored import BoredAPI
+from Core.Commands.Text.Functions.bored import BoredAPI
 
 from Core.Commands.Text.Functions.fun_fact import FunFact
 from Core.Commands.Text.Functions.covid import Covid
@@ -17,4 +17,3 @@ class Text(commands.Cog):
         self.bot.add_cog(FunFact(bot))
         self.bot.add_cog(BoredAPI(bot))
         self.bot.add_cog(Covid(bot))
-
